fix middleware.ts location
