+++ src/middleware.ts
@@ -1,0 +1,35 @@
+import { getToken } from "next-auth/jwt"
+import { withAuth } from "next-auth/middleware"
+import { NextResponse } from "next/server"
+
+export default withAuth(
+  async function middleware(req) {
+    const token = await getToken({ req })
+    const isAuth = !!token
+    const isAuthPage = req.nextUrl.pathname === "/"
+    if (isAuthPage) {
+      if (isAuth) {
+        return NextResponse.redirect(new URL("/dashboard", req.url))
+      }
+      return null
+    }
+
+    if (!isAuth) {
+      let from = req.nextUrl.pathname
+      if (req.nextUrl.search) {
+        from += req.nextUrl.search
+      }
+
+      return NextResponse.redirect(
+        new URL(`/?from=${encodeURIComponent(from)}`, req.url)
+      )
+    }
+  },
+  {
+    callbacks: {
+      async authorized() {
+        return true
+      },
+    },
+  }
+)
